requirements.txt
# --- core rtcrm ---
fastapi==0.114.0
uvicorn[standard]==0.30.6
orjson==3.10.7
python-dotenv==1.0.1

# --- google ai studio (gemini) ---
google-generativeai==0.7.0
# protobuf must be <5 for google-generativeai deps
protobuf==4.25.3

# --- typing + web stack (compatible with FastAPI 0.114) ---
pydantic==2.11.4
starlette==0.38.2

# --- optional dev/runtime helpers ---
watchfiles==1.1.0
openai>=1.30.0
httpx>=0.27.0

SQLAlchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.3
